Refactor indexeddb - remove promise hell
